requirements.txt
jax
jaxlib
matplotlib.pyplot
flax
os
hydra
